requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
pyarrow>=14.0.0
snowflake-connector-python>=3.6.0
python-dotenv>=1.0.0
dbt-snowflake>=1.7.0
streamlit>=1.30.0
plotly>=5.18.0
boto3>=1.26.0
